refactor(ai): split assistant.py into focused modules
- embedding_indexer.py: Handles codebase indexing, chunking, embedding storage
- query_processor.py: Handles query classification and model selection
- cost_tracker.py: Handles cost calculation and tracking
- chat_handler.py: Handles interactive chat sessions
- assistant.py: Main orchestrator (reduced from 795 to ~280 lines)

Benefits:
- Single Responsibility Principle: Each module has one clear purpose
- Easier to test: Components can be tested independently
- Better maintainability: Changes are isolated to relevant modules
- More extensible: Easy to add new features without touching unrelated code
